Split codes
