+++ .gitignore
@@ -1,5 +1,5 @@
-/test*.c
-/test*.exe
+# /test.c
+test*.exe
 -_*
 desktop.ini
 /.idea/
@@ -7,3 +7,4 @@ desktop.ini
 /cmake-build-debug/
 /cmake-build-release/
 /CMakeLists.txt
+.vscode/*
