+++ calc.c
@@ -1,0 +1,6 @@
+#ifndef __CALC_C__
+#define __CALC_C__
+
+// TODO: 여기에 코드 입력
+
+#endif
